refactor ColocatedImage Visualiser to vis package
